% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPlots.R
\name{arPlot}
\alias{arPlot}
\title{Abnormal Return Plot}
\usage{
arPlot(ResultParserObj, firm = NULL, window = NULL, xlab = "",
  ylab = "Abnormal Returns", alpha = 0.5, facetVar = NULL, ncol = 4,
  addAAR = F, xVar = "eventTime", yVar = "ar")
}
\arguments{
\item{firm}{set this parameter if just one firm should be plotted}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{ncol}{number of facet columns}

\item{xVar}{x variable name}

\item{yVar}{y variable name}

\item{ResultParser}{An object of class \code{ResultParser}}

\item{facet}{should each firm get its own plot (default = T)}
}
\value{
a ggplot2 object
}
