% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_objects.R
\name{CAAR Results.}
\alias{CAAR Results.}
\alias{CAAResults}
\title{An R6 object that contains CAAR results.}
\format{
\code{\link[R6]{R6Class}} object.
}
\description{
An R6 object that contains CAAR results.

An R6 object that contains CAAR results.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{caar_tbl}}{CAAR results.}

\item{\code{statistics_tbl}}{CAAR test statistic results.
Class initialization}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CAAResults-new}{\code{CAAResults$new()}}
\item \href{#method-CAAResults-print}{\code{CAAResults$print()}}
\item \href{#method-CAAResults-clone}{\code{CAAResults$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAAResults-new"></a>}}
\if{latex}{\out{\hypertarget{method-CAAResults-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAAResults$new(caar_tbl, statistics_tbl)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{caar_tbl}}{CAAR result table.}

\item{\code{statistics_tbl}}{Table with statistics.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAAResults-print"></a>}}
\if{latex}{\out{\hypertarget{method-CAAResults-print}{}}}
\subsection{Method \code{print()}}{
Print key characteristics.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAAResults$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAAResults-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CAAResults-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAAResults$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
