% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHandling.R
\name{checkClass}
\alias{checkClass}
\title{Check column type}
\usage{
checkClass(x, y, class = "integer")
}
\arguments{
\item{x}{vector}

\item{y}{column label}

\item{class}{variable type to check}
}
\description{
Function for map2 function from purrr. This function checks 
columns on the correct type
}
\keyword{internal}
