% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHandling.R
\name{checkDateFormat}
\alias{checkDateFormat}
\title{Check on correct date format}
\usage{
checkDateFormat(x)
}
\arguments{
\item{x}{a character vector}
}
\description{
The date format must be in the form %d.%m.%Y
}
\keyword{internal}
