% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requestHandling.R
\name{doHttrRequest}
\alias{doHttrRequest}
\title{Get URL content based on local}
\usage{
doHttrRequest(
  url,
  request_type = "GET",
  the_body = NULL,
  encode = "json",
  config = NULL,
  simplifyVector = getOption("EventStudy.jsonlite.simplifyVector")
)
}
\arguments{
\item{url}{the url of the page to retrieve}

\item{request_type}{the type of httr request function: GET, POST, PUT, DELETE etc.}

\item{the_body}{body of POST request}

\item{simplifyVector}{Passed to jsonlite::fromJSON}

\item{params}{A named character vector of other parameters to add to request.}

\item{customConfig}{list of httr options such as \code{httr::use_proxy}
or \code{httr::add_headers} that will be added to the request.}
}
\description{
This changes the auth type depending on if its local
}
\details{
Example of params: c(param1="foo", param2="bar")
}
\keyword{internal}
