% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{errorMessage}
\alias{errorMessage}
\title{Get the error message}
\usage{
errorMessage(x)
}
\arguments{
\item{test_me}{an object that has failed is.error}
}
\value{
The error message
}
\description{
Get the error message
}
\keyword{internal}
