% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESTARCParameters.R
\docType{data}
\name{ESTARCParameters}
\alias{ESTARCParameters}
\title{Abnormal Return Calculation Parameters}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ESTARCParameters
}
\description{
This R6 class defines the parameters for the Event Study.
}
\keyword{datasets}

