% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPerformAnalysisAddin.R
\name{estPerformAnalysisAddin}
\alias{estPerformAnalysisAddin}
\title{RStudio Addin for performing an Event Study}
\usage{
estPerformAnalysisAddin()
}
\description{
Call this as an addin to replace occurrences of a particular sequence of
characters in a document with a new sequence of characters. Te interface is 
similar to our Event Study web interface \url{www.eventstudytools.com}.
}
