% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventStudy.R
\docType{package}
\name{EventStudy}
\alias{EventStudy}
\alias{EventStudy-package}
\title{EventStudy}
\description{
This package provides functionality for doing Event Studies from R by using EventStudyTools.com API interface, parsing results, and visualize them.
}
\details{
To learn more about EventStudy visit project website: \url{www.eventstudytools.com}
or start with the vignettes:
\code{browseVignettes(package = "EventStudy")}
}
