% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventStudyApplicationInput.R
\docType{data}
\name{EventStudyApplicationInput}
\alias{EventStudyApplicationInput}
\title{Abnormal Return Calculation (ARC) API Wrapper}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EventStudyApplicationInput
}
\description{
This R6 class serialzes an ARC parameter class to a list structure
}
\keyword{datasets}
