% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{getDefaultApplicationInput}
\alias{getDefaultApplicationInput}
\title{Returns default parameters for an EventStudy type}
\usage{
getDefaultApplicationInput(type = "arc")
}
\arguments{
\item{type}{type of event study (default: arc)}
}
\description{
Creates a default parameter object for performing an Event Study
}
\keyword{internal}
