% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SP500Example.R
\name{getSP500ExampleFiles}
\alias{getSP500ExampleFiles}
\title{This function copies the 3 csv files to the actual working directory}
\usage{
getSP500ExampleFiles(targetDir = getwd())
}
\description{
This function copies the 3 csv files to the actual working directory
}
