% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPlots.R
\name{hcAarPlot}
\alias{hcAarPlot}
\title{Highchart version of Averaged Abnormale Return Plot}
\usage{
hcAarPlot(df, group = NULL, window = NULL, xlab = "",
  ylab = "Averaged Abnormal Returns")
}
\arguments{
\item{df}{data.frame with abnormal return in long format;}

\item{group}{set this parameter if just one group should be plotted}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{facet}{should each firm get its own plot (default = T)}

\item{ncol}{number of facet columns}
}
\value{
a highchart object
}
\description{
Highchart version of Averaged Abnormale Return Plot
}
