% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPlots.R
\name{hcArPlot}
\alias{hcArPlot}
\title{Highchart version of the Abnormal Return Plot}
\usage{
hcArPlot(df, firm = NULL, window = NULL, xlab = "",
  ylab = "Abnormal Returns", xVar = "eventTime", yVar = "ar")
}
\arguments{
\item{df}{data.frame with abnormal return in long format;}

\item{firm}{set this parameter if just one firm should be plotted}

\item{window}{}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{xVar}{x variable name}

\item{yVar}{y variable name}

\item{facet}{should each firm get its own plot (default = T)}

\item{ncol}{number of facet columns}
}
\value{
a highchart object
}
\description{
Highchart version of the Abnormal Return Plot
}
