% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{isError}
\alias{isError}
\title{Is this a try error?}
\usage{
isError(x)
}
\arguments{
\item{test_me}{an object created with try()}
}
\value{
Boolean
}
\description{
Utility to test errors
}
\keyword{internal}
