% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPlots.R
\name{pointwiseCARPlot}
\alias{pointwiseCARPlot}
\title{Pointwise Cumulative Abnormal Return Plot}
\usage{
pointwiseCARPlot(df, firm = NULL, xlab = "",
  ylab = "pointwise Cumulative Abnormal Returns", facet = T, ncol = 4)
}
\arguments{
\item{df}{data.frame with abnormal return in long format;}

\item{firm}{set this parameter if just one firm should be plotted}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{facet}{should each firm get its own plot (default = T)}

\item{ncol}{number of facet columns}
}
\value{
a ggplot2 object
}
