% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R, R/requestHandling.R
\name{retryRequest}
\alias{retryRequest}
\title{This functionality is borrowed from: https://github.com/MarkEdmondson1234/googleAuthR
ReTry API requests for certain errors using exponential backoff.}
\usage{
retryRequest(f)

retryRequest(f)
}
\arguments{
\item{f}{A function of a http request}

\item{f}{A function of a http request}
}
\description{
This functionality is borrowed from: https://github.com/MarkEdmondson1234/googleAuthR
ReTry API requests for certain errors using exponential backoff.

ReTry API requests for certain errors using exponential backoff.
}
\keyword{internal}
