% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPlots.R
\name{aarPlot}
\alias{aarPlot}
\title{Averaged Abnormal Return Plot}
\usage{
aarPlot(ResultParserObj, cumSum = F, group = NULL, window = NULL,
  ciStatistics = NULL, p = 0.95, ciType = "band", xlab = "",
  ylab = "Averaged Abnormal Returns", facet = T, ncol = 4)
}
\arguments{
\item{cumSum}{plot CAAR}

\item{group}{set this parameter if just one group should be plotted}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{facet}{should each firm get its own plot (default = T)}

\item{ncol}{number of facet columns}

\item{ResultParser}{An object of class \code{ResultParser}}
}
\value{
a ggplot2 object
}
