% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPlots.R
\name{aarPlot}
\alias{aarPlot}
\title{Averaged Abnormal Return Plot}
\usage{
aarPlot(df, group = NULL, window = NULL, xlab = "",
  ylab = "Averaged Abnormal Returns", facet = T, ncol = 4)
}
\arguments{
\item{df}{data.frame with abnormal return in long format;}

\item{group}{set this parameter if just one group should be plotted}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{facet}{should each firm get its own plot (default = T)}

\item{ncol}{number of facet columns}
}
\value{
a ggplot2 object
}
\description{
Averaged Abnormal Return Plot
}
