% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationInputInterface.R
\name{ApplicationInputInterface}
\alias{ApplicationInputInterface}
\title{Abstract Application Input Interface for Event Study R Package}
\format{
\code{\link[R6]{R6Class}} object.
}
\description{
Abstract class no description.
}
\section{Methods}{

\describe{
  \item{\code{$new()}}{Constructor for ApplicationInputInterface. 
  This class should not used directly.}
  \item{\code{$setNamedList()}}{Function to setup hierarchical \code{lists}.}
  \item{\code{$serialize()}}{Seriealize a \code{R6-Object}.}
  \item{\code{$serializeToJson(level)}}{Seriealize a \code{R6-Object} as a 
  \code{JSON-Object}.}
}
}

\section{Arguments}{

\describe{
  \item{\code{type}}{Set the result file type.}
  \item{\code{level}}{Level to seriealize.}
}
}

\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-setResultFileType}{\code{ApplicationInputInterface$setResultFileType()}}
\item \href{#method-setNamedList}{\code{ApplicationInputInterface$setNamedList()}}
\item \href{#method-serializeToJson}{\code{ApplicationInputInterface$serializeToJson()}}
\item \href{#method-serialize}{\code{ApplicationInputInterface$serialize()}}
\item \href{#method-clone}{\code{ApplicationInputInterface$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setResultFileType"></a>}}
\if{latex}{\out{\hypertarget{method-setResultFileType}{}}}
\subsection{Method \code{setResultFileType()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationInputInterface$setResultFileType(type = "csv")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setNamedList"></a>}}
\if{latex}{\out{\hypertarget{method-setNamedList}{}}}
\subsection{Method \code{setNamedList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationInputInterface$setNamedList(
  parentLevel,
  secondLevel = NULL,
  thirthLevel = NULL,
  parameterList
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-serializeToJson"></a>}}
\if{latex}{\out{\hypertarget{method-serializeToJson}{}}}
\subsection{Method \code{serializeToJson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationInputInterface$serializeToJson(level = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-serialize"></a>}}
\if{latex}{\out{\hypertarget{method-serialize}{}}}
\subsection{Method \code{serialize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationInputInterface$serialize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationInputInterface$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
