% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationInputInterface.R
\docType{data}
\name{ApplicationInputInterface}
\alias{ApplicationInputInterface}
\title{Abstract Application Input Interface for Event Study R Package}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApplicationInputInterface
}
\description{
Abstract class no description
}
\keyword{datasets}
