% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationInputInterface.R
\docType{data}
\name{ApplicationInputInterface}
\alias{ApplicationInputInterface}
\title{Abstract Application Input Interface for Event Study R Package}
\format{\code{\link[R6]{R6Class}} object.}
\usage{
ApplicationInputInterface
}
\description{
Abstract class no description.
}
\section{Methods}{

\describe{
  \item{\code{$new()}}{Constructor for ApplicationInputInterface. 
  This class should not used directly.}
  \item{\code{$setNamedList()}}{Function to setup hierarchical \code{lists}.}
  \item{\code{$serialize()}}{Seriealize a \code{R6-Object}.}
  \item{\code{$serializeToJson(level)}}{Seriealize a \code{R6-Object} as a 
  \code{JSON-Object}.}
}
}

\section{Arguments}{

\describe{
  \item{\code{type}}{Set the result file type.}
  \item{\code{level}}{Level to seriealize.}
}
}

\keyword{internal}
