% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArcApplicationInput.R
\docType{data}
\name{ArcApplicationInput}
\alias{ArcApplicationInput}
\title{Abnormal Return Calculation (ARC) API Wrapper}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ArcApplicationInput
}
\description{
This R6 class serialzes an ARC parameter class to a list structure
}
\keyword{datasets}
