% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPlots.R
\name{arPlot}
\alias{arPlot}
\title{Abnormal Return Plot}
\usage{
arPlot(df, firm = NULL, window = NULL, xlab = "",
  ylab = "Abnormal Returns", facet = T, ncol = 4, xVar = "eventTime",
  yVar = "ar")
}
\arguments{
\item{df}{data.frame with abnormal return in long format;}

\item{firm}{set this parameter if just one firm should be plotted}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{facet}{should each firm get its own plot (default = T)}

\item{ncol}{number of facet columns}

\item{xVar}{x variable name}

\item{yVar}{y variable name}
}
\value{
a ggplot2 object
}
\description{
Abnormal Return Plot
}
