% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPlots.R
\name{arPlot}
\alias{arPlot}
\title{Abnormal Return Plot}
\usage{
arPlot(ResultParserObj, firm = NULL, window = NULL, xlab = "",
  ylab = "Abnormal Returns", alpha = 0.5, facetVar = NULL, ncol = 4,
  addAAR = F, xVar = "eventTime", yVar = "ar")
}
\arguments{
\item{ResultParserObj}{An object of class \code{ResultParser}}

\item{firm}{set this parameter if just one firm should be plotted}

\item{window}{filter event time window}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{alpha}{alpha value}

\item{facetVar}{should each firm get its own plot (default = T)}

\item{ncol}{number of facet columns}

\item{addAAR}{add aar line}

\item{xVar}{x variable name}

\item{yVar}{y variable name}
}
\value{
a ggplot2 object
}
\description{
Abnormal return plots
}
