% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AVCApplicationInput.R
\docType{data}
\name{AVCApplicationInput}
\alias{AVCApplicationInput}
\title{Abnormal Volume Calculation Parameters}
\format{\code{\link[R6]{R6Class}} object.}
\usage{
AVCApplicationInput
}
\value{
a ESTParameters R6 object
}
\description{
This R6 class defines the parameters for the Abnormal Volume Event Study. 
We recommend to use the \code{set} functionality to setup your Event Study, 
as we check input parameters.
}
\section{Methods}{

\describe{
  \item{\code{$new()}}{Constructor for AVCApplicationInput}
  \item{\code{$setEMail(eMail)}}{Set the e-Mail address for reporting. This 
  functionality is currently not working.}
  \item{\code{$setBenchmarkModel(model = 'mm')}}{Setter for the benchmark
  model.s}
  \item{\code{$setReturnType(returnType)}}{Setter for the return type (log 
  or simple)}
  \item{\code{$setTestStatistics(testStatistics)}}{Setter for the test 
  statistics.}
}
}

\section{Arguments}{

\describe{
 \item{AVCApplicationInput}{An \code{AVCApplicationInput} object}
 \item{eMail}{An E-Mail address in \code{String} format}
 \item{model}{A benchmark model in \code{String} format}
 \item{returnType}{A return type in \code{String} format}
 \item{testStatistics}{A \code{String} vector with test statistics.}
}
}

\seealso{
\url{https://www.eventstudytools.com/axc/upload}
}
\keyword{datasets}
