% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AVCApplicationInput.R
\docType{data}
\name{AVCApplicationInput}
\alias{AVCApplicationInput}
\title{Abnormal Volume Calculation Parameters}
\format{\code{\link[R6]{R6Class}} object.}
\usage{
AVCApplicationInput
}
\value{
a ESTParameters R6 object
}
\description{
This R6 class defines the parameters for the Abnormal Volume Event Study. 
We recommend to use the \code{set} functionality to setup your Event Study, 
as we check input parameters.
}
\section{Fields}{

\describe{
\item{\code{task}}{Actually, just \code{locale} is defined. E-Mail-Address is added
in a later version.}

\item{\code{return_type}}{return type calculation: log (default), simple}

\item{\code{non_trading_days}}{handler for non-trading days: later (default), 
earlier, keep, and skip}

\item{\code{test_statistics}}{test statistics that will be used in the Event Study}

\item{\code{request_file}}{list of request file key and type}

\item{\code{firm_data}}{list of firm data file key and type}

\item{\code{market_data}}{list of market data file key and type}

\item{\code{allowedTestStatistics}}{allowed test statistics}

\item{\code{allowedBenchmarkModel}}{available market models}

\item{\code{allowedNonTradingDays}}{available market models}
}}

\section{Methods}{

\describe{
  \item{\code{$new()}}{Constructor for AVCApplicationInput}
  \item{\code{$setEMail(eMail)}}{Set the e-Mail address for reporting. This 
  functionality is currently not working.}
  \item{\code{$setBenchmarkModel(model = 'mm')}}{Setter for the benchmark
  model.s}
  \item{\code{$setReturnType(returnType)}}{Setter for the return type (log 
  or simple)}
  \item{\code{$setTestStatistics(testStatistics)}}{Setter for the test 
  statistics.}
}
}

\section{Arguments}{

\describe{
 \item{AVCApplicationInput}{An \code{AVCApplicationInput} object}
 \item{eMail}{An E-Mail address in \code{String} format}
 \item{model}{A benchmark model in \code{String} format}
 \item{returnType}{A return type in \code{String} format}
 \item{testStatistics}{A \code{String} vector with test statistics.}
}
}

\section{Class Members}{

}

\seealso{
\url{https://www.eventstudytools.com/axc/upload}
}
\keyword{datasets}
