% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AVyCApplicationInput.R
\docType{data}
\name{AVyCApplicationInput}
\alias{AVyCApplicationInput}
\title{Abnormal Volatility Calculation Parameters}
\format{\code{\link[R6]{R6Class}} object.}
\usage{
AVyCApplicationInput
}
\value{
a ESTParameters R6 object
}
\description{
This R6 class defines the parameters for the Abnormal Volatility Volume Event 
Study. We recommend to use the \code{set} functionality to setup your Event  
Study, as we check input parameters.

For more details see the help vignette:
\code{vignette("parameters_eventstudy", package = "EventStudy")}
}
\section{Methods}{


\describe{
  \item{\code{$new()}}{Constructor for AVyCApplicationInput.}
  \item{\code{$setEMail(eMail)}}{Set the e-Mail address for reporting. This 
  functionality is currently not working.}
  \item{\code{$setBenchmarkModel(model = 'mm')}}{Setter for the benchmark
  model.s}
  \item{\code{$setReturnType(returnType)}}{Setter for the return type (log 
  or simple)}
  \item{\code{$setTestStatistics(testStatistics)}}{Setter for the test 
  statistics. Per default all available test statistics are applied. You may
  find all test statistics in the vignette 'parameter_eventstudy'.}
}
}

\section{Arguments}{

\describe{
 \item{AVyCApplicationInput}{An \code{AVyCApplicationInput} object}
 \item{eMail}{An E-Mail address in \code{String} format}
 \item{model}{A benchmark model in \code{String} format}
 \item{returnType}{A return type in \code{String} format}
 \item{testStatistics}{A \code{String} vector with test statistics.}
}
}

\examples{
# get files for our S&P500 example; 3 files are written in the current 
# working directory
getSP500ExampleFiles()

# Generate a new parameter object
avycParams <- AVyCApplicationInput$new()

# set test statistics
avycParams$setTestStatistics(c("aarptlz", "aarrankz"))

# Setup API object
apiKey <- "{Your API key}"
estSetup <- EventStudyAPI$new()
estSetup$authentication(apiKey)

# Perform Event Study
estSetup$performEventStudy(estParams = avycParams, 
                           dataFiles = c("request_file" = "01_RequestFile.csv",
                                         "firm_data"    = "02_firmData.csv",
                                         "market_data"  = "03_marketData.csv"))

# Download task results and save them in the actiual working directory
estSetup$getTaskResults()

}
\seealso{
\url{https://www.eventstudytools.com/axc/upload}
}
\keyword{datasets}
