% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHandling.R
\name{checkFile}
\alias{checkFile}
\title{Chheck input data files}
\usage{
checkFile(path, type = "request_file")
}
\arguments{
\item{path}{path to the input data file}
}
\value{
data.frame
}
\description{
Check correct column, date, and shape of the input 
data files
}
