% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHandling.R
\name{checkFiles}
\alias{checkFiles}
\title{Check EventStudy input files}
\usage{
checkFiles(dataFiles = c(request_file = "01_RequestFile.csv", firm_data =
  "02_firmData.csv", market_data = "03_MarketData.csv"), returnData = F)
}
\arguments{
\item{dataFiles}{a named character vector. The names must be request_file, 
firm_data, and market_data}

\item{returnData}{returns the data as list of data.frames}
}
\description{
check each input file plus inter file relations, e.g. is market
index in request file has correct index in market_data and firm index in 
request_file has correct index in firm_data.
}
