% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESTARCParameters.R
\docType{data}
\name{ESTARCParameters}
\alias{ESTARCParameters}
\title{Get parameters for EventStudyTools API}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ESTARCParameters
}
\description{
Get parameters for EventStudyTools API
}
\keyword{datasets}

