% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESTCataParameters.R
\docType{data}
\name{ESTCataParameters}
\alias{ESTCataParameters}
\title{CATA Parameters}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ESTCataParameters
}
\keyword{datasets}
