% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventStudyAddin.R
\name{EventStudyAddin}
\alias{EventStudyAddin}
\title{RStudio Addin for performing an Event Study}
\usage{
EventStudyAddin()
}
\description{
Call this as an addin to perform an Event Study on an interface in R. The 
interface is similar to our Event Study web interface 
\url{https://www.eventstudytools.com}.
}
