% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventStudyAPI.R
\docType{data}
\name{EventStudyAPI}
\alias{EventStudyAPI}
\title{API for EventStudyTools.com}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EventStudyAPI
}
\description{
API for EventStudyTools.com
}
\keyword{datasets}

