% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventStudyApplicationInput.R
\docType{data}
\name{EventStudyApplicationInput}
\alias{EventStudyApplicationInput}
\title{Abnormal Return Calculation (ARC) API Wrapper}
\format{\code{\link{R6Class}} object.}
\usage{
EventStudyApplicationInput
}
\description{
This R6 class serialzes an Event Study parameter class to a list structure.
This is an abstract class for Event Study applications (Return, Volatility, 
and Volume Event Studies). It is not intendet to use this class directly. 
Please use: \link{ARCApplicationInput}.
}
\section{Methods}{

\describe{
  \item{\code{$new()}}{Constructor for EventStudyApplicationInput}
  \item{\code{$setup()}}{Setup the parameter list.}
}
}

\keyword{datasets}
