% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPlots.R
\name{pointwiseCARPlot}
\alias{pointwiseCARPlot}
\title{Pointwise Cumulative Abnormal Return Plot}
\usage{
pointwiseCARPlot(df, firm = NULL, xlab = "",
  ylab = "pointwise Cumulative Abnormal Returns", facetVar = NULL,
  ncol = 4)
}
\arguments{
\item{df}{data.frame with abnormal return in long format;}

\item{firm}{set this parameter if just one firm should be plotted}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{facetVar}{should each firm get its own plot. You may plot each firm in
an own plot or by each group. (Default: NULL, available: Group and Firm)}

\item{ncol}{number of facet columns}
}
\value{
a ggplot2 object
}
\description{
Pointwise cumulative abnormal return plots
}
\examples{
# plot abnormal returns in one plot
arPlot(resultParser)

# plot abnormal returns by group
arPlot(resultParser, facetVar = "Group")

This function must be revised

}
\keyword{internal}
